requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
easygoogletranslate==0.0.4
fastapi==0.105.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.2
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.1
googleapis-common-protos==1.62.0
gradio==4.9.0
gradio_client==0.7.2
greenlet==3.0.2
grpcio==1.60.0
grpcio-status==1.60.0
gTTS==2.4.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
hyperframe==5.2.0
idna==2.10
importlib-resources==6.1.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.0
langsmith==0.0.70
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
orjson==3.9.10
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
pypdf==3.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
semantic-version==2.10.0
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
tomlkit==0.12.0
toolz==0.12.0
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
websockets==11.0.3
wheel==0.41.2
yarl==1.9.4
